backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-simple-history==3.4.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
python-decouple==3.8
dj-database-url==2.1.0
django-filter==23.5
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.2.0

